MenuItem 置き場を集約する
