+++ Assets/UniGLTF/Editor/Menu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 727dafd648a17b74c951a2f08187637e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
