+++ Assets/UniGLTF/Editor/UniGLTF/Serialization/DeserializerGenerator.cs
@@ -39,8 +39,7 @@ public static class GltfDeserializer
             }
         }
 
-        [MenuItem(UniGLTFVersion.MENU + "/GLTF: Generate Deserializer")]
-        static void GenerateSerializer()
+        public static void GenerateSerializer()
         {
             var info = new ObjectSerialization(typeof(glTF), "gltf", "Deserialize_");
             Debug.Log(info);
